с помощью fetch получаем данные для галлерии загруженных фото
