+++ index.html
@@ -239,7 +239,7 @@
     <script src="vendor/nouislider/nouislider.js"></script>
     <script src="./js/functions.js" type="module"></script>
     <script src="./js/main.js" type="module"></script>
-    <script src="./js/create-pictures.js" type="module"></script>
+    <script src="./js/create-gallery.js" type="module"></script>
     <script src="./js/render-big-picture.js" type="module"></script>
     <script src="./js/picture-modal.js" type="module"> </script>
     <script src="./js/form.js" type="module"></script>
