+++ js/main.js
@@ -1,4 +1,10 @@
-import { gallery } from './gallery.js';
+//import { gallery } from './gallery.js';
+import { createGallery } from './create-gallery.js';
+import './picture-modal.js';
+import './form.js';
 
-
-JSON.stringify(gallery, null, 2);
+fetch('https://31.javascript.htmlacademy.pro/kekstagram/data')
+  .then((response) => response.json())
+  .then((photos) => {
+    createGallery(photos);
+  });
